\name{bmt}
\alias{bmt}
\docType{data}
\title{
Bone marrow transplant study
}
\description{
Bone marrow transplant study used for modeling AFT semicure model}
\usage{data(bmt)}
\format{
  A data frame with 91 observations on the following 3 variables.
  \describe{
    \item{\code{Time}}{a numeric vector}
    \item{\code{Status}}{a numeric vector}
    \item{\code{TRT}}{a numeric vector}
  }
}
\examples{
data(bmt)
## maybe str(bmt) ; plot(bmt) ...
}
\keyword{datasets}
