\name{coefsmcure}
\alias{coefsmcure}
\title{
retrieves the estimated coefficients from object
}
\description{
Retrieving coefficients, printing method, and summary mthod for a smcure object. 
}
\usage{
coefsmcure(x, ...)
}
\arguments{
  \item{x}{
a smcure object}
 \item{\dots}{
Further arguments to be passed to the print.smcure function.
}
}
