\name{e1684}
\alias{e1684}
\docType{data}
\title{
Eastern Cooperative Oncology Group (ECOG) data
}
\description{
Eastern Cooperative Oncology Group (ECOG) data used for modeling PH semicure model}
\usage{data(e1684)}
\format{
  A data frame with 285 observations on the following 5 variables.
  \describe{
    \item{\code{TRT}}{a numeric vector}
    \item{\code{FAILTIME}}{a numeric vector}
    \item{\code{FAILCENS}}{a numeric vector}
    \item{\code{AGE}}{a numeric vector}
    \item{\code{SEX}}{a numeric vector}
  }
}

\examples{
data(e1684)
## maybe str(e1684) ; plot(e1684) ...
}
\keyword{datasets}
