\name{predictsmcure}
\alias{predictsmcure}
\title{
prediction of semicure model}
\description{
prediction of semicure model}
\usage{
predictsmcure(object, newX, newZ, model = c("ph", "aft"), ...)
}
\arguments{
  \item{object}{
object of smcure
}
  \item{newX}{
New value(s) of X}
  \item{newZ}{
New value(s) of Z}
  \item{model}{
either 'ph' or 'aft'}
 \item{\dots}{
Further arguments to be passed to the predictsmcure function}
}