\name{smcure-package}
\alias{smcure-package}
\alias{smcure-package}
\docType{package}
\title{
Semiparametric mixture cure model
}
\description{
An R package to fit semiparametric PH mixture cure model or AFT mixture cure model
}
\details{
\tabular{ll}{
Package: \tab smcure\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2012-02-05\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Chao Cai, Yubo Zou, Yingwei Peng, Jiajia Zhang
Maintainer: Chao Cai <caic@email.sc.edu>
}

\references{
Kersey, et al., Comparison of autologous and allogeneic bone marrow trans-plantation for treatment of high-risk refractory acute lymphoblastic leukemia. New England Journal of Medicine, 317 (8), 461-467, 1987.


Kirkwood, et al., Interferon alfa-2b adjuvant therapy of high-risk resected cutaneous melanoma: the eastern cooperative oncology group trial est 1684. Journal of Clinical Oncology, 14 (1), 7, 1996.


J. Zhang and Y. Peng. A new estimation method for the semiparametric accelerated
failure time mixture cure model. Statistics in Medicine, 26(16):3157-3171, 2007. 


Peng, Y., Dear, K. B. G., 2000. A nonparametric mixture model for cure rate estimation. Biometrics
56 (1), 237-243. 


C. Cai, et al., smcure: An R-Package for estimating semiparametric mixture cure models, Computer Methods and Programs in Biomedicine. (2012), http://dx.doi.org/10.1016/j.cmpb.2012.08.013.


}

