\name{bmt}
\alias{bmt}
\docType{data}
\title{
Bone marrow transplant study
}
\description{
Bone marrow transplant study for the refractory acute lymphoblastic leukemia patients which is widely used in the AFTMC model (Kersey, et al., 1987)}
\usage{data(bmt)}
\format{
 There were 90 patients in total, where 46 patients in the allogeneic treatment group and 44 patients in the autologous treatment group
  \describe{
    \item{\code{Time}}{ 
time to death
}
    \item{\code{Status}}{
censoring indicator, 1=event of interest happens, and 0=censoring
}
    \item{\code{TRT}}{
1 for autologous treatment group; 0 for allogeneic treatment group
}
  }
}



\keyword{datasets}
