\name{coefsmcure}
\alias{coefsmcure}
\title{
Retrieves the estimated coefficients from object
}
\description{
Retrieving coefficients from a smcure object. 
}
\usage{
coefsmcure(x,...) 
}
\arguments{
  \item{x}{
an object from smcure function
}
 \item{\dots}{
further arguments to be passed to the \code{printsmcure} function.
}
}
